Set up robinhood util file for hashes, reductions, transforms, metadata definitions, etc.
