+++ inc/zoo/map/RobinHood.h
@@ -3,6 +3,7 @@
 
 #include "zoo/swar/SWAR.h"
 #include "zoo/AlignedStorage.h"
+#include "zoo/map/RobinHoodUtil.h"
 
 #ifndef ZOO_CONFIG_DEEP_ASSERTIONS
     #define ZOO_CONFIG_DEEP_ASSERTIONS 0
@@ -18,88 +19,8 @@
 
 namespace zoo {
 
-template<typename T>
-struct GeneratorFromPointer {
-    T *p_;
-
-    constexpr T &operator*() noexcept { return *p_; }
-    constexpr GeneratorFromPointer operator++() noexcept { return {++p_}; }
-};
-
-template<typename T, int MisalignmentBits>
-struct MisalignedGenerator {
-    T *base_;
-    constexpr static auto Width = sizeof(T) * 8;
-
-    constexpr T operator*() noexcept {
-        auto firstPart = base_[0];
-        auto secondPart = base_[1];
-            // how to make sure the "logical" shift right is used, regardless
-            // of the signedness of T?
-            // I'd prefer to not use std::make_unsigned_t, since how do we
-            // "make unsigned" user types?
-        auto firstPartLowered = firstPart.value() >> MisalignmentBits;
-        auto secondPartRaised = secondPart.value() << (Width - MisalignmentBits);
-        return T{firstPartLowered | secondPartRaised};
-    }
-
-    constexpr MisalignedGenerator operator++() noexcept { return {++base_}; }
-};
-
-template<typename T>
-struct MisalignedGenerator<T, 0>: GeneratorFromPointer<T> {};
-
-template<typename T>
-struct MisalignedGenerator_Dynamic {
-    constexpr static auto Width = sizeof(T) * 8;
-    T *base_;
-
-    int misalignmentFirst, misalignmentSecond;
-
-    MisalignedGenerator_Dynamic(T *base, int ma):
-        base_(base),
-        misalignmentFirst(ma), misalignmentSecond(Width - ma)
-    {}
-    
-
-    constexpr T operator*() noexcept {
-        auto firstPart = base_[0];
-        auto secondPart = base_[1];
-            // how to make sure the "logical" shift right is used, regardless
-            // of the signedness of T?
-            // I'd prefer to not use std::make_unsigned_t, since how do we
-            // "make unsigned" user types?
-        auto firstPartLowered = firstPart.value() >> misalignmentFirst;
-        auto secondPartRaised = secondPart.value() << misalignmentSecond;
-        return T{firstPartLowered | secondPartRaised};
-    }
-
-    constexpr MisalignedGenerator_Dynamic operator++() noexcept {
-        ++base_; return *this;
-    }
-};
 
 namespace rh {
-
-namespace impl {
-
-/// \todo decide on whether to rename this?
-template<int PSL_Bits, int HashBits, typename U = std::uint64_t>
-struct Metadata: swar::SWARWithSubLanes<HashBits, PSL_Bits, U> {
-    using Base = swar::SWARWithSubLanes<HashBits, PSL_Bits, U>;
-    using Base::Base;
-
-    constexpr auto PSLs() const noexcept { return this->least(); }
-    constexpr auto Hashes() const noexcept { return this->most(); }
-};
-
-template<int PSL_Bits, int HashBits, typename U = std::uint64_t>
-struct MatchResult {
-    U deadline;
-    Metadata<PSL_Bits, HashBits, U> potentialMatches;
-};
-
-}
 
 template<int PSL_Bits, int HashBits, typename U = std::uint64_t>
 struct RH_Backend {
@@ -289,41 +210,6 @@ struct RH_Backend {
         }
     }
 };
-
-template<int NBits, typename U>
-constexpr auto hashReducer(U n) noexcept {
-    constexpr auto Shift = (NBits * ((sizeof(U)*8 / NBits) - 1));
-    constexpr auto AllOnes = meta::BitmaskMaker<U, 1, NBits>::value;
-    auto temporary = AllOnes * n;
-    auto higestNBits = temporary >> Shift;
-    return
-        (0 == (64 % NBits)) ?
-            higestNBits :
-            swar::isolate<NBits>(higestNBits);
-}
-
-
-template<typename T>
-constexpr auto fibonacciIndexModulo(T index) {
-    constexpr std::array<uint64_t, 4>
-        GoldenRatioReciprocals = {
-            159,
-            40503,
-            2654435769,
-            11400714819323198485ull,
-        };
-    constexpr T MagicalConstant =
-        T(GoldenRatioReciprocals[meta::logFloor(sizeof(T)) - 3]);
-    return index * MagicalConstant;
-}
-
-template<size_t Size, typename T>
-constexpr auto lemireModuloReductionAlternative(T input) noexcept {
-    static_assert(sizeof(T) == sizeof(uint64_t));
-    static_assert(Size < (1ull << 32));
-    auto halved = uint32_t(input);
-    return Size * input >> 32;
-}
 
 template<
     typename K,
